Fixed GridDB mapping 2 queries

+++ src/main/java/edu/uci/ics/tippers/schema/griddb/mappings/GridDBSchemaMapping2.java
@@ -178,7 +178,7 @@ public class GridDBSchemaMapping2 extends GridDBBaseSchemaMapping {
         columnInfoList.add(new ColumnInfo("timeStamp", GSType.TIMESTAMP));
         columnInfoList.add(new ColumnInfo("id", GSType.STRING, indexSet));
         columnInfoList.add(new ColumnInfo("sensorId", GSType.STRING));
-        columnInfoList.add(new ColumnInfo("temperature", GSType.DOUBLE));
+        columnInfoList.add(new ColumnInfo("temperature", GSType.INTEGER));
 
         containerInfo.setColumnInfoList(columnInfoList);
         containerInfo.setRowKeyAssigned(true);
